Adding monitoring button on sources page and Delete Alert Confirm

+++ src/Components/SourceItem/index.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import { Button } from "react-bootstrap";
 import moment from "moment";
-import { FiTrash } from "react-icons/fi";
+import { FiTrash, FiClock } from "react-icons/fi";
+import { useHistory } from "react-router-dom";
 
 import { apiCrawlers } from "../../services/api";
 
@@ -42,16 +43,26 @@ interface SourceItemProps {
 }
 
 const SourceItem: React.FC<SourceItemProps> = ({ item, onDelete, onClick }) => {
+  
+  const history = useHistory(); 
+  
   const deleteSource = async (e: React.SyntheticEvent) => {
     e.stopPropagation();
-    await apiCrawlers
-      .delete(`crawlers/${item.id}/`)
-      .then(() => {
-        onDelete();
-      })
-      .catch(() => {
-        alert("Ocorreu um erro inesperado ao deletar a fonte!");
-      });
+    if (confirm('Tem certeza que deseja deletar essa fonte?')) {
+      await apiCrawlers
+        .delete(`crawlers/${item.id}/`)
+        .then(() => {
+          onDelete();
+        })
+        .catch(() => {
+          alert("Ocorreu um erro inesperado ao deletar a fonte!");
+        });
+    }
+  };
+
+  const monitoringSource = (e: React.SyntheticEvent) => {
+    e.stopPropagation();
+    history.push(`/fontes/monitoramento/${item.id}`); 
   };
 
   return (
@@ -61,6 +72,9 @@ const SourceItem: React.FC<SourceItemProps> = ({ item, onDelete, onClick }) => {
         <ResultDate>
           {moment(item.created_at).format("DD/MM/YYYY hh:mm")}
         </ResultDate>
+        <Button onClick={monitoringSource}>
+          <FiClock />
+        </Button>
         <Button onClick={deleteSource}>
           <FiTrash />
         </Button>
